requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
bottle==0.12.25
bottle-websocket==0.2.9
bs4==0.0.1
certifi==2023.5.7
charset-normalizer==3.2.0
colorama==0.4.6
contourpy==1.2.0
crispy==0.7.4
crispy-bootstrap4==2023.1
cycler==0.12.1
Django==4.2.3
django-ckeditor==6.7.0
django-cleanup==8.0.0
django-crispy-forms==2.1
django-extensions==3.2.3
django-filter==23.3
django-js-asset==2.1.0
django-phonenumber-field==7.2.0
django-resized==1.0.2
django-rest-passwordreset==1.3.0
djangorestframework==3.14.0
drf-yasg==1.21.7
Eel==0.16.0
fabio==2023.10.0
fonttools==4.44.0
future==0.18.3
geojson==2.5.0
gevent==23.7.0
gevent-websocket==0.10.1
greenlet==2.0.2
h5py==3.10.0
hdf5plugin==4.3.0
idna==3.4
inflection==0.5.1
kiwisolver==1.4.5
lxml==4.9.3
matplotlib==3.8.1
numpy==1.26.1
packaging==23.2
phonenumbers==8.13.22
Pillow==10.0.0
psycopg2-binary==2.9.9
pyowm==3.3.0
pyparsing==3.1.0
PyQt5==5.15.10
PyQt5-Qt5==5.15.11
PyQt5-sip==12.13.0
PySocks==1.7.1
pyTelegramBotAPI==4.12.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
silx==1.1.2
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
telebot==0.0.5
uritemplate==4.1.1
urllib3==2.0.3
whichcraft==0.6.1
zope.event==5.0
zope.interface==6.0
